+++ src/entrypoint.sh
@@ -3,7 +3,10 @@
 # https://demyx.sh
 set -euo pipefail
 
-# Generate config
+# Generate demyx config
+[[ ! -d "$NGINX_CONFIG"/nginx ]] && tar -xzf /etc/demyx.tgz -C "$NGINX_CONFIG"
+
+# Generate wp config
 if [[ "$WORDPRESS" = true ]]; then
     [[ ! -f "$NGINX_CONFIG"/nginx.conf ]] && demyx-wp
 else